Feat: ReservationState is now determined using 'Hotelier' or 'Stay' instead of 'App'
